int -> long

+++ pingpongcallc/c_src/ping.c
@@ -3,14 +3,14 @@
 
 int p = 0;
 
-void user_do_ping_spg(int *v) {
+void user_do_ping_spg(long long *v) {
     printf("*** SENDING PING *** %d\n", p);
     *v = p;
     p++;
     fflush(stdout);
 }
 
-void user_ping_spg(int i) {
+void user_ping_spg(long long i) {
     printf("*** PING *** %d\n", i);
     fflush(stdout);
 }
